container_prototype/requirements.txt
botorch
tensorflow
lume-model @ git+https://github.com/slaclab/lume-model.git@2921e6583a6cfd49285833eb851b361aacf65b4c
pandas>=2.0.1
numpy>=1.21.2
keras==3.1.1
mlflow==2.11.3
boto3==1.34.65
psycopg2-binary
k2eg==0.2.5




